lots of renaming to better differentiate the simulator and the encoder
