+++ bliss/datasets/galsim_galaxies.py
@@ -11,7 +11,7 @@ from bliss.catalog import FullCatalog
 from bliss.datasets.background import ConstantBackground
 from bliss.datasets.sdss import convert_flux_to_mag
 from bliss.metrics import get_single_galaxy_ellipticities
-from bliss.models.galsim_decoder import (
+from bliss.simulator.galsim_decoder import (
     DefaultGalsimPrior,
     FullCatalogDecoder,
     SingleGalsimGalaxyDecoder,
