+++ bliss/detection_encoder.py
