+++ bliss/inference.py
@@ -13,7 +13,7 @@ from bliss.datasets.sdss import SloanDigitalSkySurvey
 from bliss.datasets.simulated import SimulatedDataset
 from bliss.encoder import Encoder
 from bliss.metrics import CoaddFullCatalog
-from bliss.models.decoder import ImageDecoder
+from bliss.simulator.decoder import ImageDecoder
 
 
 def reconstruct_scene_at_coordinates(
