+++ case_studies/starnet/plotting.py
@@ -1,7 +1,7 @@
 import torch
 from matplotlib import pyplot as plt
 
-from bliss.models.detection_encoder import get_is_on_from_n_sources
+from bliss.detection_encoder import get_is_on_from_n_sources
 
 
 def plot_image(axarr, image, x0=0, x1=0, slen0=100, slen1=100):
