+++ case_studies/summer_template/config.yaml
@@ -20,7 +20,7 @@ datasets:
     sdss_galaxies:
         _target_: bliss.datasets.galsim_galaxies.SingleGalsimGalaxies
         prior:
-            _target_: bliss.models.galsim_decoder.SingleGalsimGalaxyPrior
+            _target_: bliss.simulator.galsim_decoder.SingleGalsimGalaxyPrior
             flux_sample: "pareto"
             min_flux: 622.0
             max_flux: 1e6
@@ -31,7 +31,7 @@ datasets:
             a_scale: 4.432725319432478
             a_bulge_disk_ratio: 2.0
         decoder:
-            _target_: bliss.models.galsim_decoder.SingleGalsimGalaxyDecoder
+            _target_: bliss.simulator.galsim_decoder.SingleGalsimGalaxyDecoder
             n_bands: 1
             slen: 53
             pixel_scale: 0.396
@@ -51,8 +51,8 @@ datasets:
         valid_n_batches: 10  #200
     simulated:
         _target_: bliss.datasets.simulated.SimulatedDataset
-        decoder: ${models.decoder}
-        prior: ${models.prior}
+        decoder: ${simulator.decoder}
+        prior: ${simulator.prior}
         n_tiles_h: 8
         n_tiles_w: 8
         background:
@@ -69,9 +69,9 @@ datasets:
         fix_validation_set: true
         valid_n_batches: 10  #256
 
-models:
+simulator:
     prior:
-        _target_: bliss.models.prior.ImagePrior
+        _target_: bliss.simulator.prior.ImagePrior
         n_bands: 1
         max_sources: 1
         mean_sources: 0.2
@@ -81,14 +81,14 @@ models:
         alpha: 0.43
         prob_galaxy: 0.30
         galaxy_prior:
-            _target_: bliss.models.prior.GalaxyPrior
+            _target_: bliss.simulator.prior.GalaxyPrior
             n_latent_batches: 160
-            autoencoder: ${models.galaxy_net}
+            autoencoder: ${simulator.galaxy_net}
             autoencoder_ckpt: ${paths.project}/models/sdss_autoencoder.pt
             latents_file: ${paths.project}/models/latents_simulated_sdss_galaxies.pt
             galaxy_dataset: ${datasets.sdss_galaxies}
     galaxy_net:
-        _target_: bliss.models.galaxy_net.OneCenteredGalaxyAE
+        _target_: bliss.simulator.galaxy_net.OneCenteredGalaxyAE
         slen: 53
         latent_dim: 8
         hidden: 256
@@ -96,35 +96,36 @@ models:
         optimizer_params:
             lr: 1e-3
     decoder:
-        _target_: bliss.models.decoder.ImageDecoder
+        _target_: bliss.simulator.decoder.ImageDecoder
         n_bands: 1
         tile_slen: 4
         ptile_slen: 52
         border_padding: 24
         psf_params_file: ${paths.data}/sdss/94/1/12/psField-000094-1-0012.fits
         galaxy_model:
-            _target_: bliss.models.galaxy_net.OneCenteredGalaxyAE
-            slen: ${models.galaxy_net.slen}
-            latent_dim: ${models.galaxy_net.latent_dim}
-            hidden: ${models.galaxy_net.hidden}
-            n_bands: ${models.galaxy_net.n_bands}
+            _target_: bliss.simulator.galaxy_net.OneCenteredGalaxyAE
+            slen: ${simulator.galaxy_net.slen}
+            latent_dim: ${simulator.galaxy_net.latent_dim}
+            hidden: ${simulator.galaxy_net.hidden}
+            n_bands: ${simulator.galaxy_net.n_bands}
             ckpt: ${paths.project}/models/sdss_autoencoder.pt
         psf_slen: 25
         sdss_bands:
             - 2
-    detection_encoder:
-        _target_: bliss.models.detection_encoder.DetectionEncoder
-        n_bands: ${models.decoder.n_bands}
-        tile_slen: ${models.decoder.tile_slen}
-        tiles_to_crop: 6
-        annotate_probs: True
-        slack: 1.0
-        optimizer_params:
-            lr: 1e-3
-        scheduler_params:
-            milestones: [32]
-            gamma: 0.1
-        architecture: ${like_yolov5l}
+
+encoder:
+    _target_: bliss.detection_encoder.DetectionEncoder
+    n_bands: ${simulator.decoder.n_bands}
+    tile_slen: ${simulator.decoder.tile_slen}
+    tiles_to_crop: 6
+    annotate_probs: True
+    slack: 1.0
+    optimizer_params:
+        lr: 1e-3
+    scheduler_params:
+        milestones: [32]
+        gamma: 0.1
+    architecture: ${like_yolov5l}
 
 # this architecture is based on yolov5l.yaml, see
 # https://github.com/ultralytics/yolov5/blob/master/models/yolov5l.yaml
@@ -166,7 +167,7 @@ like_yolov5l:
 training:
     name: "sdss_detection_encoder"
     dataset: ${datasets.simulated}
-    model: ${models.detection_encoder}
+    encoder: ${encoder}
     version: null
     n_epochs: 1000
     experiment: default
